requirements.txt
pytest==8.2.2
requests==2.32.3
selenium==4.22.0
allure-pytest==2.13.5
webdriver-manager==4.0.2